Bugfix: cancel on "new passphrase" dialog now works on Centos.
The previous commit worked fine on Debian Squeeze, but on a Centos CVL
desktop, hitting Cancel on the "new passphrase" dialog box resulted in
a spinning busy mouse cursor and no dialog saying "did not install key".

Turned out to be due to a missing 'else' case in getNewPassphrase_stage1().